requirements.txt
google-cloud-bigquery>=3.17.2
dify-plugin-sdk>=0.3.0 